// src/shared/services/api/models/userDetailDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { RefTermDto } from './refTermDto';

export interface UserDetailDto {
  id?: string;
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  mobile_number?: string | null;
  designation?: RefTermDto;
  gender?: RefTermDto;
  current_tier?: RefTermDto;
  /** @nullable */
  roles?: string[] | null;
  date_created?: string;
  /** @nullable */
  profile_image?: string | null;
}