// src/shared/services/api/models/userSummaryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */

export interface UserSummaryDto {
  id?: string;
  /** @nullable */
  full_name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  roles?: string[] | null;
}